moved galleries to their own folder; made getPage function
