+++ pages/travel/uzbekistan/index.tsx
@@ -3,24 +3,7 @@ import KhivaImage from '@public/images/uzbekistan-khiva-night.jpg'
 import SamarkandImage from '@public/images/uzbekistan-samarkand-dome.jpg'
 import Layout from '@src/components/common/Layout'
 import TravelLinkList from '@src/components/site/TravelLinkList'
-import { CountryGallery } from '@src/photos/utils/types'
-
-export const countryGallery: CountryGallery = {
-    country: 'uzbekistan',
-    cloudinaryFolder: 'Uzbekistan\\ 2023/Samarkand',
-    title: 'Uzbekistan',
-    keywords: [
-        'Uzbekistan',
-        'Tashkent',
-        'Bukhara',
-        'Samarkand',
-        'Khiva',
-        'Asia',
-        'Central Asia',
-        'Silk road',
-        'travel',
-    ],
-}
+import countryGallery from '../../../src/galleries/uzbekistan'
 
 export default function Page() {
     return (
