+++ pages/travel/uzbekistan/photos/[[...photos]].tsx
@@ -1,5 +1,6 @@
-import { countryGallery } from '@pages/travel/uzbekistan'
 import Layout from '@src/components/common/Layout'
+import cityGalleryMap from '@src/galleries/uzbekistan/samarkand'
+import countryGallery from '@src/galleries/uzbekistan'
 import { GalleryIndexPage } from '@src/photos/components/GalleryPage'
 import { SinglePhotoPage } from '@src/photos/components/SinglePhotoPage'
 import getGalleryStaticPaths from '@src/photos/pageHelpers/getGalleryStaticPaths'
@@ -13,18 +14,6 @@ import { useRouter } from 'next/router'
 export type PageProps = {
     images: ImageProps[]
     currentPhoto?: ImageProps | null
-}
-
-const samarkandGalleryMap: CityGallery = {
-    country: 'uzbekistan',
-    city: 'samarkand',
-    cloudinaryFolder: 'Uzbekistan\\ 2023/Samarkand',
-    title: 'Samarkand, Uzbekistan',
-    keywords: ['Uzbekistan', 'Samarkand', 'Asia', 'Central Asia', 'Silk road', 'travel'],
-}
-
-const cityGalleryMap: CityGalleryMap = {
-    samarkand: samarkandGalleryMap,
 }
 
 function GalleryPage({
@@ -58,56 +47,60 @@ function GalleryPage({
     )
 }
 
-export default function Page({ images, currentPhoto }: PageProps) {
-    const router = useRouter()
-    const photosParam = router.query.photos
-    const segments: string[] = Array.isArray(photosParam) ? photosParam : photosParam ? [photosParam] : []
+function getPage(countryGallery: CountryGallery, cityGalleryMap: CityGalleryMap) {
+    return function Page({ images, currentPhoto }: PageProps) {
+        const router = useRouter()
+        const photosParam = router.query.photos
+        const segments: string[] = Array.isArray(photosParam) ? photosParam : photosParam ? [photosParam] : []
 
-    if (segments && segments.length > 0) {
-        const cityName = segments[0]
+        if (segments && segments.length > 0) {
+            const cityName = segments[0]
 
-        if (!cityName || !(cityName in cityGalleryMap)) {
-            throw new Error(`invalid city name in query params: ${cityName}`)
-        }
-
-        const cityGallery = cityGalleryMap[cityName]
-
-        // Index page for a specific city
-        if (segments.length === 1) {
-            return <GalleryPage gallery={cityGallery} cityGalleryMap={cityGalleryMap} images={images} />
-        }
-
-        // Page for a specific photo
-        if (segments.length === 2) {
-            if (!currentPhoto) {
-                // This will happen when loading an index page and then clicking on a photo
-                const photoIdFromProps = getPhotoIdFromRouter(router.query)
-                currentPhoto = images.find((img) => img.id === photoIdFromProps)
+            if (!cityName || !(cityName in cityGalleryMap)) {
+                throw new Error(`invalid city name in query params: ${cityName}`)
             }
 
-            if (!currentPhoto) {
-                throw new Error('missing the currentPhoto prop')
+            const cityGallery = cityGalleryMap[cityName]
+
+            // Index page for a specific city
+            if (segments.length === 1) {
+                return <GalleryPage gallery={cityGallery} cityGalleryMap={cityGalleryMap} images={images} />
             }
 
-            return (
-                <Layout title={countryGallery.title} keywords={cityGallery.keywords}>
-                    <h1>{countryGallery.title}</h1>
+            // Page for a specific photo
+            if (segments.length === 2) {
+                if (!currentPhoto) {
+                    // This will happen when loading an index page and then clicking on a photo
+                    const photoIdFromProps = getPhotoIdFromRouter(router.query)
+                    currentPhoto = images.find((img) => img.id === photoIdFromProps)
+                }
 
-                    <p>
-                        <Link href={`/travel/${countryGallery.country}`}>{`Back to ${countryGallery.title}`}</Link>
-                    </p>
+                if (!currentPhoto) {
+                    throw new Error('missing the currentPhoto prop')
+                }
 
-                    <SinglePhotoPage cityGallery={cityGallery} currentPhoto={currentPhoto} images={images} />
-                </Layout>
-            )
+                return (
+                    <Layout title={countryGallery.title} keywords={cityGallery.keywords}>
+                        <h1>{countryGallery.title}</h1>
+
+                        <p>
+                            <Link href={`/travel/${countryGallery.country}`}>{`Back to ${countryGallery.title}`}</Link>
+                        </p>
+
+                        <SinglePhotoPage cityGallery={cityGallery} currentPhoto={currentPhoto} images={images} />
+                    </Layout>
+                )
+            }
+
+            console.error('cannot figure out where to go')
         }
 
-        console.error('cannot figure out where to go')
+        // Index page for all of the country's photos
+        return <GalleryPage gallery={countryGallery} cityGalleryMap={cityGalleryMap} images={images} />
     }
-
-    // Index page for all of the country's photos
-    return <GalleryPage gallery={countryGallery} cityGalleryMap={cityGalleryMap} images={images} />
 }
+
+export default getPage(countryGallery, cityGalleryMap)
 
 export const getStaticProps: GetStaticProps = async function (context) {
     return getGalleryStaticProps(Object.values(cityGalleryMap), context)
