+++ src/galleries/uzbekistan/index.ts
@@ -1,0 +1,20 @@
+import { CountryGallery } from '@src/photos/utils/types'
+
+const countryGallery: CountryGallery = {
+    country: 'uzbekistan',
+    cloudinaryFolder: 'Uzbekistan\\ 2023/Samarkand',
+    title: 'Uzbekistan',
+    keywords: [
+        'Uzbekistan',
+        'Tashkent',
+        'Bukhara',
+        'Samarkand',
+        'Khiva',
+        'Asia',
+        'Central Asia',
+        'Silk road',
+        'travel',
+    ],
+}
+
+export default countryGallery
