+++ src/galleries/uzbekistan/samarkand.ts
@@ -1,0 +1,13 @@
+import type { CityGalleryMap } from '@src/photos/utils/types'
+
+const cityGalleryMap: CityGalleryMap = {
+    samarkand: {
+        country: 'uzbekistan',
+        city: 'samarkand',
+        cloudinaryFolder: 'Uzbekistan\\ 2023/Samarkand',
+        title: 'Samarkand, Uzbekistan',
+        keywords: ['Uzbekistan', 'Samarkand', 'Asia', 'Central Asia', 'Silk road', 'travel'],
+    },
+}
+
+export default cityGalleryMap
